# backend/openapi/mlms.openapi.yaml
openapi: 3.0.3
info:
  title: MLMS API
  version: 1.0.0
  description: |
    Medical Laboratory Management System API.

    Supports the complete laboratory workflow:
    - Patient registration
    - Order creation
    - Specimen collection
    - Result entry
    - Report validation and publication
    - Patient portal access

    All protected endpoints require a Bearer JWT token obtained via `POST /auth/login`.
servers:
  - url: http://localhost:3000/api
    description: Local development server
...
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /specimens:
    post:
      tags: [Specimens]
      summary: Register a specimen
      description: Requires TECHNICIAN or ADMIN role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpecimenRequestDto'
            example:
              orderId: clxyz002
              barcode: LAB-2026-00001
              type: BLOOD
              collectionTime: '2026-03-16T09:00:00.000Z'
              notes: null
      responses:
        '201':
          description: Specimen created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecimenDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Barcode already exists
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /specimens/{id}:
    get:
      tags: [Specimens]
      summary: Get specimen by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Specimen found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecimenDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /specimens/{id}/status:
    patch:
      tags: [Specimens]
      summary: Update specimen status
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSpecimenStatusRequestDto'
            example:
              status: RECEIVED
              receivedAt: '2026-03-16T09:15:00.000Z'
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecimenDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /specimens/{id}/results:
    get:
      tags: [Results]
      summary: Get all results for a specimen
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Results for specimen
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResultDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ── Results ────────────────────────────────────────────────────────────────
  /results:
    post:
      tags: [Results]
      summary: Record a laboratory result
      description: Requires TECHNICIAN or ADMIN role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordResultRequestDto'
            example:
              specimenId: clxyz003
              testCode: GLU
              testNameFr: Glycémie
              testNameAr: سكر الدم
              value: '7.2'
              unit: mmol/L
              referenceLow: 3.9
              referenceHigh: 6.1
              flag: H
              measuredAt: '2026-03-16T10:00:00.000Z'
      responses:
        '201':
          description: Result recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Result for this testCode already exists on this specimen
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /results/{id}:
    patch:
      tags: [Results]
      summary: Update a result
      description: Requires TECHNICIAN or ADMIN role.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResultRequestDto'
      responses:
        '200':
          description: Result updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ── Reports ────────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Parameters ──────────────────────────────────────────────────────────────
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: Resource ID (cuid)
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          example:
            error:
              code: UNAUTHORIZED
              message: Authentication required
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          example:
            error:
              code: FORBIDDEN
              message: Access denied
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
    ConflictError:
      description: Conflict – resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          example:
            error:
              code: CONFLICT
              message: Resource already exists
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          example:
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                - field: firstName
                  message: firstName is required

  # ── Schemas ─────────────────────────────────────────────────────────────────
  schemas:
...
    SpecimenType:
      type: string
      enum: [BLOOD, URINE, STOOL, TISSUE]
    SpecimenStatus:
      type: string
      enum: [COLLECTED, RECEIVED, PROCESSED, DISPOSED, REJECTED]
    ResultFlag:
      type: string
      enum: [N, H, L, HH, LL, CRITICAL]
...
    ApiFieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          example: firstName
        message:
          type: string
          example: firstName is required
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            $ref: '#/components/schemas/ApiFieldError'
    ApiErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'

    # ── Pagination ─────────────────────────────────────────────────────────
...
    SpecimenDto:
      type: object
      required: [id, orderId, barcode, type, status, collectionTime, createdAt, updatedAt]
      properties:
        id:
          type: string
        orderId:
          type: string
        barcode:
          type: string
          example: LAB-2026-00001
        type:
          $ref: '#/components/schemas/SpecimenType'
        status:
          $ref: '#/components/schemas/SpecimenStatus'
        collectionTime:
          type: string
          format: date-time
        receivedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSpecimenRequestDto:
      type: object
      required: [orderId, barcode, type, collectionTime]
      properties:
        orderId:
          type: string
        barcode:
          type: string
        type:
          $ref: '#/components/schemas/SpecimenType'
        collectionTime:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
    UpdateSpecimenStatusRequestDto:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/SpecimenStatus'
        receivedAt:
          type: string
          format: date-time
          nullable: true

    # ── Result ─────────────────────────────────────────────────────────────
    ResultDto:
      type: object
      required: [id, specimenId, testCode, testNameFr, testNameAr, value, flag, measuredAt, createdAt, updatedAt]
      properties:
        id:
          type: string
        specimenId:
          type: string
        testCode:
          type: string
          example: GLU
        testNameFr:
          type: string
          example: "Glycémie"
        testNameAr:
          type: string
          example: "سكر الدم"
        value:
          type: string
          example: '7.2'
        unit:
          type: string
          nullable: true
          example: mmol/L
        referenceLow:
          type: number
          nullable: true
          example: 3.9
        referenceHigh:
          type: number
          nullable: true
          example: 6.1
        flag:
          $ref: '#/components/schemas/ResultFlag'
        measuredAt:
          type: string
          format: date-time
        measuredBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RecordResultRequestDto:
      type: object
      required: [specimenId, testCode, testNameFr, testNameAr, value, flag, measuredAt]
      properties:
        specimenId:
          type: string
        testCode:
          type: string
        testNameFr:
          type: string
        testNameAr:
          type: string
        value:
          type: string
        unit:
          type: string
          nullable: true
        referenceLow:
          type: number
          nullable: true
        referenceHigh:
          type: number
          nullable: true
        flag:
          $ref: '#/components/schemas/ResultFlag'
        measuredAt:
          type: string
          format: date-time
    UpdateResultRequestDto:
      type: object
      properties:
        value:
          type: string
        unit:
          type: string
          nullable: true
        referenceLow:
          type: number
          nullable: true
        referenceHigh:
          type: number
          nullable: true
        flag:
          $ref: '#/components/schemas/ResultFlag'
        measuredAt:
          type: string
          format: date-time

    # ── Report ─────────────────────────────────────────────────────────────
...